requirements.txt
Pillow==9.5.0
os-sys==2.1.4